// src/app/shared/model/anomalie.model.ts
export class AnomalieModel {
  idAnomalie: number;
  idFrontPointage: string;
  idEmployee: number;
  idRestaurant: number;
  idFront: string;
  badgeEmployee: string ;
  nomEmploye: string ;
  isPreAlarme: boolean;
  prenomEmploye: string ;
  valeurDepasse: string;
  valeurContrainte: string;
  libelleAnomalie: string;
  codeAnomalie: string;
  valide: boolean;
  dateAnomalie: string | Date;

}